Aun no funciona al 100%, pero he hecho algunos cambios en APPConsola para que permita seguir leyendo de los Json y en teoría tiene la estructura CRUD para la bd SQL. Sin embargo, me queda re-adaptar los métodos en los ficheros "repository" para que encajen con la nueva lógica de AppConsola.

+++ src/main/java/com/web/maven/bancoABMModel/model/CuentaBancaria.java
@@ -17,7 +17,7 @@ public class CuentaBancaria {
         this.movimientos = new ArrayList<>();
     }
 
-    public CuentaBancaria(String numeroCuenta, BigDecimal saldoInicial) {
+    public CuentaBancaria(String numeroCuenta, double saldoInicial) {
         this.numeroCuenta = numeroCuenta;
         this.saldo = saldoInicial != null ? saldoInicial : BigDecimal.ZERO;
         this.movimientos = new ArrayList<>();
@@ -31,7 +31,7 @@ public class CuentaBancaria {
         this.numeroCuenta = numeroCuenta;
     }
 
-    public BigDecimal getSaldo() {
+    public double getSaldo() {
         return saldo;
     }
